+++ lib/items_selector.dart
@@ -5,7 +5,11 @@ export 'src/model/select_options.dart';
 export 'src/model/wrap_configuration.dart';
 export 'src/model/list_configuration.dart' hide BaseListConfiguration;
 export 'src/model/grid_configuration.dart' hide BaseGridConfiguration;
+export 'src/model/radio_selector_item.dart';
+export 'src/model/check_box_selector_item.dart';
 
 export 'src/widget/grid_selector.dart' hide GridSelectorState;
 export 'src/widget/list_selector.dart' hide ListSelectorState, ListSelectorType;
 export 'src/widget/wrap_selector.dart' hide WrapSelectorState;
+export 'src/widget/radio_selector.dart';
+export 'src/widget/check_box_selector.dart';